+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Added
 - Start of `quickstart` commands
+- Added `tdx.converters.NicknameConverter`
 
 ### Changed
 - Moved translations to a submodule for Crowdin support